requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cachetools==3.1.0
cffi==1.12.2
Click==7.0
cryptography==2.6.1
Flask==1.0.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
paramiko==2.4.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyNaCl==1.3.0
rsa==4.0
six==1.12.0
sshtunnel==0.1.4
uritemplate==3.0.0
Werkzeug==0.15.2
